major reorganisation, but pytest green
