--- scopesim_templates/basic/__init__.py
@@ -1,7 +1,0 @@
-from . import stars
-from . import galaxy
-from . import misc
-
-
-from .basic import source_from_image, source_from_file
-
