--- scopesim_templates/basic/misc.py
@@ -1,25 +1,0 @@
-import numpy as np
-
-from ..rc import Source, __config__
-from ..utils.general_utils import function_call_str
-
-
-def empty_sky():
-    """
-    Returns an empty source so that instrumental fluxes can be simulated
-
-    Returns
-    -------
-    sky : Source
-
-    """
-    params = {"function_call": function_call_str(empty_sky, {}),
-              "object": "empty sky"}
-
-    sky = Source(lam=np.array([__config__["!spectral.wave_min"],
-                               __config__["!spectral.wave_max"]]),
-                 spectra=np.array([0, 0]), x=[0], y=[0], ref=[0], weight=[0])
-    sky.meta.update(params)
-
-    return sky
-
