+++ scopesim_templates/calibration/__init__.py
@@ -1,0 +1,2 @@
+from .calibration import *
+from . import micado
