+++ scopesim_templates/calibration/micado/flatlamp.py
