--- scopesim_templates/exgal/__init__.py
@@ -1,2 +1,0 @@
-from ..advanced.galaxy import galaxy, galaxy3d
-from ..basic.galaxy import elliptical, spiral_two_component
