+++ scopesim_templates/extragalactic/__init__.py
@@ -1,0 +1,4 @@
+from .galaxies import *
+from .clusters import *
+from .agns import *
+from .quasars import *
