+++ scopesim_templates/extragalactic/exgal_models.py
