+++ scopesim_templates/extragalactic/galaxies.py
@@ -1,20 +1,22 @@
 import numpy as np
-import synphot
+import pyckles
 
-from astropy.io import fits
-from astropy.utils.data import download_file
+import synphot
 from astropy import units as u
-from astropy.utils.decorators import deprecated_renamed_argument
+from astropy.io import fits
+from astropy.utils import deprecated_renamed_argument
+from astropy.utils.data import download_file
 from astropy.wcs import WCS
 
 from spextra import Spextrum
-from scopesim.source.source_templates import Source
 
-from .. import rc
+from ..rc import Source, __config__
+from ..rc import im_plane_utils as ipu
+from ..rc import ter_curve_utils as tcu
+from ..extragalactic import galaxy_utils as gal_utils
+from ..extragalactic.exgal_models import GalaxyBase
+from ..misc.misc import source_from_image
 from ..utils import general_utils as gu
-from ..basic.basic import source_from_image
-
-from ..utils.exgal_models import GalaxyBase
 
 
 @deprecated_renamed_argument('plate_scale', 'pixel_scale', '0.1')
@@ -228,7 +230,6 @@ def galaxy3d(sed,           # The SED of the galaxy,
 
     return src
 
-#-------------------------------------------------
 
 def spiral_two_component(extent=60*u.arcsec, fluxes=(0, 0), offset=(0, 0)):
     """
@@ -265,8 +266,8 @@ def spiral_two_component(extent=60*u.arcsec, fluxes=(0, 0), offset=(0, 0)):
         extent /= 3600.
 
     filename = "spiral_two_component.fits"
-    url = rc.__config__["!file.server_url"]
-    use_cache = rc.__config__["!file.use_cache"]
+    url = __config__["!file.server_url"]
+    use_cache = __config__["!file.use_cache"]
     print(url+filename)
 
     path = download_file(remote_url=url+filename, cache=use_cache)
@@ -274,7 +275,7 @@ def spiral_two_component(extent=60*u.arcsec, fluxes=(0, 0), offset=(0, 0)):
     img_ext = hdulist[0].header["IMG_EXT"]
     spec_ext = hdulist[0].header["SPEC_EXT"]
 
-    src = rc.Source()
+    src = Source()
     src.fields = hdulist[img_ext:spec_ext]
     src.spectra = [gu.hdu_to_synphot(hdu) for hdu in hdulist[spec_ext:]]
 
@@ -300,78 +301,130 @@ def spiral_two_component(extent=60*u.arcsec, fluxes=(0, 0), offset=(0, 0)):
 
     return src
 
-#-------------------------------------------------
 
-def spiral_two_component(extent=60*u.arcsec, fluxes=(0, 0), offset=(0, 0)):
+@deprecated_renamed_argument('magnitude', 'amplitude', '0.1')
+def elliptical(half_light_radius, pixel_scale, filter_name, amplitude,
+               spectrum="NGC_0584", **kwargs):
     """
-    Creates a spiral galaxy using NGC1232L as the template
+    Create a extended :class:`.Source` object for an elliptical galaxy
 
-    Two components are included
-        - the newer population (spiral arms), and
-        - the older popultaion (bulge)
+    .. note:: This docstring is from simcado, needs to be updated
 
     Parameters
     ----------
-    extent : float
+    half_light_radius : float
         [arcsec]
-    fluxes : list of floats, Quantity
-        [mag | ABmag | jansky | FLAM | FNU | PHOTLAM | PHOTNU]
-    offset : tuple of floats
+
+    pixel_scale : float
         [arcsec]
+
+    amplitude : float
+        [mag, mag/arcsec2]
+
+    n : float, optional
+        Power law index. Default = 4
+        - n=1 for exponential (spiral),
+        - n=4 for de Vaucouleurs (elliptical)
+
+    filter_name : str, TransmissionCurve, optional
+        Default is "Ks". Values can be either:
+        - the name of a SimCADO filter : see optics.get_filter_set()
+        - or a TransmissionCurve containing a user-defined filter
+
+    normalization : str, optional
+        ["half-light", "centre", "total"] Where the profile equals unity
+        If normalization equals:
+        - "half-light" : the pixels at the half-light radius have a surface
+                         brightness of ``magnitude`` [mag/arcsec2]
+        - "centre" : the maximum pixels have a surface brightness of
+                     ``magnitude`` [mag/arcsec2]
+        - "total" : the whole image has a brightness of ``magnitude`` [mag]
+
+    spectrum : str, optional
+        The spectrum to be associated with the galaxy. Values can either be:
+        - the name of a SimCADO SED spectrum : see get_SED_names()
+        - an EmissionCurve with a user defined spectrum
+
+
+    Optional Parameters (passed to ``sersic_profile``)
+    --------------------------------------------------
+    ellipticity : float
+        Default = 0.5
+
+    angle : float
+        [deg] Default = 30. Rotation anti-clockwise from the x-axis
+
+    width, height : int
+        [arcsec] Dimensions of the image. Default: 512*pixel_scale
+
+    x_offset, y_offset : float
+        [arcsec] The distance between the centre of the profile and the centre
+        of the image. Default: (dx,dy) = (0,0)
+
 
     Returns
     -------
-    gal : scopesim.Source
+    galaxy_src : Source
+
+
+    See Also
+    --------
 
     """
-    params = {"extent": extent,
-              "fluxes": fluxes,
-              "offset": offset}
-    pass
-    params["function_call"] = gu.function_call_str(spiral_two_component, params)
-    params["object"] = "two component spiral galaxy"
 
-    if isinstance(extent, u.Quantity):
-        if extent.unit.physical_type == "angle":
-            extent = extent.to(u.deg).value
-        else:
-            raise ValueError("Physical type of extent must be 'angle': "
-                             "".format(extent.unit.physical_type))
-    else:
-        extent /= 3600.
+    # """
+    # 1 make a sersic profile ImageHDU
+    # 2 get spectrum from Brown
+    # 3 scale spectrum to amplitude, assume ABmag if no unit
+    # 4 make Source object
+    # """
 
-    filename = "spiral_two_component.fits"
-    url = rc.__config__["!file.server_url"]
-    use_cache = rc.__config__["!file.use_cache"]
-    print(url+filename)
+    params = {"n": 4,
+              "ellipticity": 0.5,
+              "angle": 30,
+              "normalization": "total",
+              "width": pixel_scale * 512,
+              "height": pixel_scale * 512,
+              "x_offset": 0,
+              "y_offset": 0,
+              "redshift": 0,
+              "half_light_radius": half_light_radius,
+              "pixel_scale": pixel_scale,
+              "filter_name": filter_name,
+              "amplitude": amplitude,
+              "spectrum_name": str(spectrum)}
+    params.update(kwargs)
+    params["function_call"] = gu.function_call_str(elliptical, params)
+    params["object"] = "elliptical galaxy"
 
-    path = download_file(remote_url=url+filename, cache=use_cache)
-    hdulist = fits.open(path)
-    img_ext = hdulist[0].header["IMG_EXT"]
-    spec_ext = hdulist[0].header["SPEC_EXT"]
+    # 1 make a sersic profile ImageHDU
+    im = gal_utils.sersic_profile(r_eff=half_light_radius / pixel_scale,    # everything in terms of pixels
+                                  n=params["n"],
+                                  ellipticity=params["ellipticity"],
+                                  angle=params["angle"],
+                                  normalization=params["normalization"],
+                                  width=params["width"] / pixel_scale,
+                                  height=params["height"] / pixel_scale)
 
-    src = rc.Source()
-    src.fields = hdulist[img_ext:spec_ext]
-    src.spectra = [gu.hdu_to_synphot(hdu) for hdu in hdulist[spec_ext:]]
+    hw, hh = 0.5 * params["width"], 0.5 * params["height"]
+    xs = (np.array([-hw, hw]) + params["x_offset"]) / 3600.
+    ys = (np.array([-hh, hh]) + params["y_offset"]) / 3600.
+    hdr = ipu.header_from_list_of_xy(xs, ys, pixel_scale / 3600.)
+    hdu = fits.ImageHDU(data=im, header=hdr)
 
-    for ii in range(len(src.fields)):
-        w, h = src.fields[ii].data.shape
-        src.fields[ii].header["CRPIX1"] = w // 2
-        src.fields[ii].header["CRPIX2"] = h // 2
-        src.fields[ii].header["CRVAL1"] = 0
-        src.fields[ii].header["CRVAL2"] = 0
-        src.fields[ii].header["CDELT1"] = extent / w
-        src.fields[ii].header["CDELT2"] = extent / w
-        src.fields[ii].header["CUNIT1"] = "DEG"
-        src.fields[ii].header["CUNIT2"] = "DEG"
-        src.fields[ii].header["CTYPE1"] = "RA---TAN"
-        src.fields[ii].header["CTYPE2"] = "DEC--TAN"
-        src.fields[ii].header["SPEC_REF"] = src.fields[ii].header["SPEC_EXT"] \
-                                            - spec_ext
+    # 2 get spectrum from Brown
+    if isinstance(spectrum, str):
+        brown_lib = pyckles.SpectralLibrary("brown", return_style="synphot")
+        spectrum = brown_lib[spectrum]
+        spectrum.z = params["redshift"]
 
-    # ..todo: scale image plane according to fluxes
-    # ..todo: shift header values according to offset
+    # 3 scale the spectra and get the weights
+    if not isinstance(amplitude, u.Quantity):
+        amplitude = amplitude << u.ABmag
+    spectrum = tcu.scale_spectrum(spectrum, filter_name, amplitude)
 
+    # 4 make Source object
+    src = Source(spectra=[spectrum], image_hdu=hdu)
     src.meta.update(params)
 
     return src
