+++ scopesim_templates/extragalactic/galaxy_utils.py
