+++ scopesim_templates/stellar/__init__.py
@@ -1,3 +1,2 @@
-from ..basic.stars import star, star_grid, star_field
-from ..basic.stars import cluster as star_cluster
-
+from .clusters import *
+from .stars import *
