+++ scopesim_templates/stellar/imf.py
