+++ scopesim_templates/stellar/mamajek.dat
