+++ scopesim_templates/stellar/mamajek_alt.dat
