+++ scopesim_templates/stellar/stars_utils.py
