--- scopesim_templates/tests/test_advanced/test_galaxy.py
@@ -1,8 +1,0 @@
-# We need tests here
-
-
-class TestVelField:
-    def test_1(self):
-        pass
-
-
