+++ scopesim_templates/utils/__init__.py
@@ -1,4 +1,2 @@
-from . import imf
-from . import cluster_utils
-from . import stars_utils
 from . import general_utils
+from . import samplers
