+++ scopesim_templates/utils/samplers.py
@@ -1,6 +1,4 @@
 # a collection of samplers for generating distributions
-
-
 import numpy as np
