+++ scopesim_templates/version.py
@@ -1,2 +1,2 @@
-version = '0.2.2'
-date    = '2021-06-10 12:00:00 GMT'
+version = '0.3.0'
+date    = '2021-11-01 12:00:00 GMT'
